backend/requirements.txt
# Core framework
django==4.2.29
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
# ASGI / server
uvicorn==0.25.0
watchfiles>=0.21
gunicorn==25.1.0
# Database
psycopg2-binary==2.9.11
dj-database-url==3.1.2
# Static files
whitenoise==6.6.0
# PDF parsing
pdfminer.six==20221105
PyPDF2==3.0.1
Pillow==10.3.0
# Config
python-decouple==3.8
# HTTP (for LMStudio agent calls)
requests>=2.31.0
# Testing
pytest>=8.0.0
pytest-django>=4.8.0
# Dev tools
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
# Legacy – kept so existing env doesn't break
emergentintegrations==0.1.0
